requirements.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.2
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
databases==0.7.0
fastapi==0.89.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
magic-filter==1.0.9
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
psycopg2==2.9.5
py-postgresql==1.3.0
pydantic==1.10.4
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
sniffio==1.3.0
SQLAlchemy==2.0.1
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.2
